% Generated by roxygen2 (4.0.0): do not edit by hand
\name{adj2edge}
\alias{adj2edge}
\title{Convert an adjacency matrix to an edge list}
\usage{
adj2edge(adj, noEdge = c(0, NA), diag = FALSE)
}
\arguments{
\item{adj}{Adjacency matrix.}

\item{noEdge}{values in the adjacency matrix to consider as "no edge".
Defaults to 0 and NA.}

\item{diag}{Logical; Include the diagonals?}
}
\description{
Convert an adjacency matrix to an edge list
}

