% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_expectedInf.R
\name{bridge_expectedInf}
\alias{bridge_expectedInf}
\title{Bridge Expected Influence}
\usage{
bridge_expectedInf(network, communities = NULL, useCommunities = "all",
  directed = NULL, nodes = NULL, step = c("both", 1, 2))
}
\arguments{
\item{network}{an object of type \code{qgraph}, \code{igraph}, or an adjacency
matrix representing a network. Adjacency matrices should be complete (e.g., not
only upper or lower half)}

\item{communities}{an object of class "communities" (igraph) OR a characcter vector of
community  assignments for each node. The ordering of this vector should correspond
to the vector from argument "nodes"}

\item{useCommunities}{character vector specifying which communities should be included}

\item{directed}{logical. Directedness is automatically detected if set to "NULL" (the default).
Symmetric adjacency matrices will be undirected, unsymmetric matrices will be directed}

\item{nodes}{a vector containing the names of the nodes. If set to "NULL", this vector will
be automatically detected in the order given in "network"}

\item{step}{compute 1-step expected influence, 2-step expected influence,
or both}
}
\description{
Calculates the expected influence of a node on communities other than
its own community. Note that currently, 2-step bridge EI only works for graphs
with exactly 2 communities
}
\details{
When a network contains both positive and negative edges, traditional centrality measures
such as strength centrality may not accurately predict node influence on the network.
Robinaugh, Millner, & McNally (2016) showed that in these cases,
expected influence is a more appropriate measure.

One-step expected influence is defined as the sum of all edges extending
from a given node (where the sign of each edge is maintained).

Two-step expected influence, as the name implies, measures connectivity up to two edges away from the node.
It is defined as the sum of the (weighted) expected influences of each node connected to the initial node
plus the one-step expected influence of the initial node. Weights are determined by the edge strength between
the initial node and each "second step" node.

See citations in the references section for further details.
}
\references{
Robinaugh, D. J., Millner, A. J., & McNally, R. J. (2016). Identifying highly influential nodes in the complicated grief network. \emph{Journal of abnormal psychology}, 125, 747.
}
