% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{coerce_to_adjacency}
\alias{coerce_to_adjacency}
\title{Coerce to adjacency matrix}
\usage{
coerce_to_adjacency(input, directed = NULL)
}
\arguments{
\item{input}{a network of class "igraph", "qgraph", or an adjacency matrix representing
a network}

\item{directed}{logical. is the network directed? If set to NULL, auto-detection is used}
}
\description{
Takes an object of type "qgraph", "igraph", or an adjacency matrix (or data.frame) and outputs an adjacency matrix
}
