% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compareFiltered}
\alias{compareFiltered}
\title{Compare two filtered edge tables for differences}
\usage{
compareFiltered(filtered1, filtered2)
}
\arguments{
\item{filtered1}{filtered table of edges.}

\item{filtered2}{filtered table of edges.}
}
\value{
A list with the following elements:
  \itemize{
   \item{only.in.first}{A data frame of the edges only in \code{filtered1}}.
   \item{only.in.second}{A data frame of the edges only in \code{filtered2}}.
   \item{conflicting}{
     A data frame of edges whose direction is opposite between
     \code{filtered1} and \code{filtered2}.
   }
  }
}
\description{
Provides a list of edges that are only present in one filtered table, along
with a list of edges whose direction conflicts between both.
}

