% Generated by roxygen2 (4.0.0): do not edit by hand
\name{degree}
\alias{degree}
\title{Calculate the degree of each node in the network}
\usage{
degree(adjacency)
}
\arguments{
\item{adjacency}{an adjacency matrix}
}
\value{
a named vector
}
\description{
Calculate the degree of each node in the network
}

