% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{depression}
\alias{depression}
\title{Simulated Depression Profiles}
\format{
a dataframe. Columns represent symptoms and rows represent individuals
}
\usage{
depression
}
\description{
This simulated dataset contains severity ratings for 9 symptoms of major depressive
disorder in 1000 individuals. Symptom ratings are  assumed to be self-reported
on a 100 point sliding scale.
}
\examples{
head(depression)
}
\keyword{datasets}
