% Generated by roxygen2 (4.0.0): do not edit by hand
\name{estimateCausalEffectSize}
\alias{estimateCausalEffectSize}
\title{Estimate the effect size of a causal edge.}
\usage{
estimateCausalEffectSize(filtered, datC)
}
\arguments{
\item{filtered}{A table of significant edges to estimate the effect size for.}

\item{datC}{A \code{data.frame} containing the observations for each SNP and
trait.}
}
\value{
Returns the \code{data.frame} given in to the \code{filtered} parameter with
two additional columns: one for the estimated effect size, and one with the
p-value.
}
\description{
Uses the Two-stage least squares method to estimate the causal effect size
of each Source of each Target using the CPA SNPs.
}
\references{
\enumerate{
 \item{
   Angrist, J. D. & Imbens, G. W. Two-Stage Least Squares Estimation of
   Average Causal Effects in Models with Variable Treatment Intensity.
   J. Am. Stat. Assoc. 90, 431–442 (1995).
 }
 \item{
   Lawlor, D. A., Harbord, R. M., Sterne, J. A. C., Timpson, N. & Davey
   Smith, G. Mendelian randomization: using genes as instruments for making
   causal inferences in epidemiology. Stat. Med. 27, 1133–1163 (2008).
 }
}
}

