% Generated by roxygen2 (4.0.0): do not edit by hand
\name{evaluateAnchors}
\alias{evaluateAnchors}
\title{Evalute the Quality of the Causal and Orthoganl Anchors chosen by NEO.}
\usage{
evaluateAnchors(edge.table, datC)
}
\arguments{
\item{dt}{edge table to be written out by \code{\link{parseNEOcsv}}.}

\item{datC}{matrix given to NEO}

\item{qtls}{a combined table of QTL associations generated by \code{plink}
for each trait. Table should have a column named \code{PHEN}}
}
\value{
\code{edge.table} with the following columns added:
  \itemize{
    \item{CPA.P}{
     Model p-value for a linear model of the CPA anchors on the Source node,
     in the subset of individuals in \code{datC}.
    }
    \item{CPA.R2}{
     Adjusted R-squared for a linear model of the CPA anchors on the Source
     node, in the subset of individuals in \code{datC}.
    }
    \item{OCA.P}{
     Model p-value for a linear model of the OCA anchors on the Target node,
     in the subset of individuals in \code{datC}.
    }
    \item{OCA.R2}{
     Adjusted R-squared for a linear model of the OCA anchors on the Target
     node, in the subset of individuals in \code{datC}.
    }
  }
}
\description{
Evalute the Quality of the Causal and Orthoganl Anchors chosen by NEO.
}

