% Generated by roxygen2 (4.0.0): do not edit by hand
\name{filterNEO}
\alias{filterNEO}
\title{Filter a NEO edge table}
\usage{
filterNEO(edge.table, min.OCA = 0.3, min.CPA = 0, min.model.P = 0.05,
  max.path.P = 0.05, min.BLV = 0, max.RMSEA = 0.05,
  max.CPA.Source.P = 5e-08, max.CPA.Target.P = 0.05,
  max.OCA.Target.P = 5e-08, min.OCA.Source.P = 0.05)
}
\arguments{
\item{edge.table}{edge.table generated by \code{\link{parseNEOcsv}}.}

\item{min.OCA}{Primary recommended threshold for significant edges by NEO (1).}

\item{min.CPA}{The OCA score is more reliable than the CPA score, but they
should be in agreement (1).}

\item{min.model.P}{Each model is assessed by a partial correlation: the
closer to 0, the more support for that model.
\code{P.Model.AB} is the p-value for the partial
correlation for the causal model A --> B, as a consistency
check, this should be non-significant.}

\item{max.path.P}{P-value from the standardised Wald Path Coefficient, which
supports the existence of an edge (1). This should be
significant.}

\item{min.BLV}{The Bi-layer V score measures the presence of confounding
paths. Recommended to be above 0 (2).}

\item{max.RMSEA}{The RMSEA is a structural edge model (sem) model-fitting
indice (See the \code{\link{sem}} package \code{NEO}
depends on). This should be below 0.05 (1).}

\item{max.CPA.Source.P}{Significance threshold for CPA SNPs on the Source
node.}

\item{max.CPA.Target.P}{Significance threshold for CPA SNPs on the Target
node. Note: We expect this p-value to be much higher
than the \code{CPA.Source.P}.}

\item{max.OCA.Target.P}{Significance threshold for the OCA SNPs on the Target
node.}

\item{min.OCA.Source.P}{The relationship between the OCA SNPs and the Source
node should be non-significant if the Source is
causal of the Target.}
}
\value{
a subset of the edge table containing only significant edges.
}
\description{
Uses a combination of recommended thresholds, along with other consistency
checks, to confidently determine which edges are significant.
}
\details{
{
NEO uses partial correlation to assess the probability of each model between
the Source and Target nodes. The closer the partial correlation for a model
is to 0, the more evidence there is for that model (1). However, the partial
correlation can also be close to 0 when the CPA and OCA SNPs NEO chooses are
not truly causal of their respective traits. Although one would expect the
\code{LEO.NB.OCA} score to be close to 0, we find that many edges will pass
the recommended threshold even when the SNPs are not at all causal of the
traits (NEO simply choses the best SNPs, but imposes no threshold on what
"best" means. In other words, Garbage-in, garbage-out). Therefore, we impose
additional consistency checks on the relationship between the causal anchors
NEO choses and their trait, based on the fundamental literature on
Instrumental Variable analysis utilised in the context of genetics
(Mendelian Randomisation) (3):

If the Source is truly causal of the Target, we expect the CPA SNPs to be
significantly associated with the Source and Target nodes (the relationship
with the Target node should be entirely mediated by the Source node, and thus
weaker). We also expect to see the OCA SNPs to NOT be significantly
associated with the Source node; the flow of information should prevent any
correlation (unless they are pleiotropicly associated with both nodes, in
which case we are less confident in the results, but you may wish to relax
this constraint) (3, 4).

If the CPA SNPs are not significantly associated with the Source node OR the
OCA SNPs are not significantly associated with the Target node (or both), the
result should be discarded completely, and you should review your pre-NEO
SNP selection and genotype loading code. In the case only one set of SNPs is
significant, the Next Best score will be unreliable, because NEO will be
biased towards the models which use the non-significant SNPs.
}
}
\references{
\enumerate{
 \item{
   Aten, J. E., Fuller, T. F., Lusis, A. J. & Horvath, S. Using genetic
   markers to orient the edges in quantitative trait networks: the NEO
   software. BMC Syst. Biol. 2, 34 (2008).
 }
 \item{
   Aten, J. E. Causal not Confounded Gene Networks: Inferring Acyclic and
   Non-acyclic Gene Bayesian Networks in mRNA Expression Studies. (2008).
   at <http://labs.genetics.ucla.edu/horvath/aten/NEO/jason_e_aten_dissertation2008.pdf>
 }
 \item{
   Lawlor, D. A., Harbord, R. M., Sterne, J. A. C., Timpson, N. & Davey
   Smith, G. Mendelian randomization: using genes as instruments for making
   causal inferences in epidemiology. Stat. Med. 27, 1133–1163 (2008).
 }
 \item{
   Pearl, J. Causality: models, reasoning and inference.
   (Cambridge Univ Press, 2000).
 }
}
}

