% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getNonOrientableEdges}
\alias{getNonOrientableEdges}
\title{Get a list of relationships that NEO was unable to orient.}
\usage{
getNonOrientableEdges(edge.table)
}
\arguments{
\item{edge.table}{The full edge table output by \code{\link{parseNEOcsv}}.}
}
\value{
an edge.table of non-orientable edges.
}
\description{
In some cases, NEO is unable to make an attempt at orienting an edge.
These edges will be missing completely from the edge.table, but it is useful
to know which edges could not be determined.
}

