% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getOrphans}
\alias{getOrphans}
\title{Identify Nodes for whom no edges could be confidently called by NEO}
\usage{
getOrphans(edge.table, filtered)
}
\arguments{
\item{edge.table}{The full edge.table}

\item{filtered}{The edge.table filtered by \code{\link{filterNEO}}.}
}
\value{
a list of nodes with no edges
}
\description{
Identify Nodes for whom no edges could be confidently called by NEO
}

