% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSnet.R
\name{MDSnet}
\alias{MDSnet}
\title{MDSnet}
\usage{
MDSnet(qgraph_net, type = c("ordinal", "interval", "ratio", "mspline"),
  MDSadj = NULL, stressTxt = F, repulse = F, repulsion = 1,
  mdsArgs = list(), ...)
}
\arguments{
\item{qgraph_net}{an object of type \code{qgraph}}

\item{type}{transformation function for MDS, defaults to "ordinal"}

\item{MDSadj}{to use a proximities matrix other than the adjacency matrix
stored in \code{qgraph_net}, provide it in this argument}

\item{stressTxt}{logical. Print the stress value in the lower left corner of the plot?}

\item{repulse}{logical. Add a small repulsion force with wordcloud package to avoid node overlap?}

\item{repulsion}{scalar for the repulsion force. Larger values add more repulsion}

\item{mdsArgs}{additional arguments in list format passed to \code{smacof::mds}}

\item{...}{additional arguments passed to \code{qgraph}}
}
\description{
Convenience function for converting a qgraph object to a layout determined
by multidimensional scaling
}
\details{
A network plotted with multidimensional scaling can be interpreted based on the distances
between nodes. Nodes close together represent closely associated nodes, whereas nodes that are far
apart represent unassociated or negatively associated nodes.
}
\references{
Jones, P. J., Mair, P., & McNally, R. J. (2018). Visualizing psychological networks: A tutorial in R. Frontiers in Psychology, 9, 1742. https://doi.org/10.3389/fpsyg.2018.01742
}
