% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mediators}
\alias{mediators}
\title{Identify mediator nodes in the network}
\usage{
mediators(adjacency)
}
\arguments{
\item{adjacency}{an adjacency matrix}
}
\value{
a vector of node names
}
\description{
Identify mediator nodes in the network
}

