% Generated by roxygen2 (4.0.0): do not edit by hand
\name{neighborhood}
\alias{neighborhood}
\title{Retrieve all nodes in the Neighborhood of a given node}
\usage{
neighborhood(node, adjacency, order, mode = "all")
}
\arguments{
\item{node}{Character constant, Node you want to retrieve neighborhood of.}

\item{adjacency}{an adjacency matrix}

\item{order}{Integer giving the order of the neighborhood.}

\item{mode}{Character constant, it specifies how to use the direction of
the edges if a directed graph is analyzed. For ‘out’ only the
outgoing edges are followed, so all vertices reachable from
the source vertex in at most ‘order’ steps are counted. For
‘"in"’ all vertices from which the source vertex is reachable
in at most ‘order’ steps are counted. ‘"all"’ ignores the
direction of the edges.}
}
\value{
a vector of node names in the requested neighborhood.
}
\description{
This function mimics the behaviour of \code{neighborhood} in the
\code{igraph} package, but is backwards compatible with \code{R 2.5.1}
}
\details{
The neighborhood of a given order ‘o’ of a vertex ‘v’ includes all
vertices which are closer to ‘v’ than the order. Ie. order 0 is
always ‘v’ itself, order 1 is ‘v’ plus its immediate neighbors,
order 2 is order 1 plus the immediate neighbors of the vertices in
order 1, etc.
}

