% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{networkTools-package}
\alias{networkTools}
\alias{networkTools-package}
\title{Tools for analysing and visualising networks.}
\description{
This R package provides functions for analysisng and visualising large
networks, primarily those generated from the Horvath et al suite of network
packages and tools, e.g. WGCNA and the NEO software
\link{http://labs.genetics.ucla.edu/horvath/aten/NEO/}.

This package is backwards compatible with older versions of R (i.e. 2.5.1).
}
\details{
\tabular{ll}{
 Package: \tab networkTools\cr
 Type: \tab Package\cr
 Version: \tab 1.0\cr
 Date: \tab 2014-02-19\cr
 License: \tab GPL (>= 2)\cr
}
}
\author{
Scott Ritchie \email{sritchie73@gmail.com}
}

