% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktools-package.r
\docType{package}
\name{networktools}
\alias{networktools}
\alias{networktools-package}
\title{networktools.}
\description{
Tools for Identifying Important Nodes in Networks
}
\details{
Includes assorted tools for network analysis. Specifically, includes functions for
calculating impact statistics, which aim to identify how each node impacts
the overall network structure (global strength impact, network structure impact, edge impact),
and for calculating and visualizing expected influence.

For a complete list of functions, use library(help = "networktools")

For a complete list of vignettes, use browseVignettes("networktools")
}
\author{
Payton J. Jones
}
