% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parseAnchors}
\alias{parseAnchors}
\title{Parse an anchor column to obtain all unique SNPs}
\usage{
parseAnchors(anchors)
}
\arguments{
\item{anchors}{the CPA.Anchors or OCA.Anchors column to parse}
}
\value{
a list of all SNPs that appear in the column
}
\description{
Parse an anchor column to obtain all unique SNPs
}

