% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parseNEOcsv}
\alias{parseNEOcsv}
\title{Converts the output of \code{NEO} into something sane.}
\usage{
parseNEOcsv(neofilename, outfile, datC)
}
\arguments{
\item{neofilename}{path to the neo csv file.}

\item{outfile}{file prefix to output the .edge.table.tsv and .snp.table.tsv
files}

\item{datC}{data matrix given to NEO as its \code{datC} argument}
}
\description{
\code{NEO} outputs a massive CSV file that is not easily parseable, nor
loadable into cytoscape. This function will parse this CSV file and write
two tables to file in tab separatted format:
An edge table, filled with relevant columns, notably
splitting out the \code{edge} column into Source and Target nodes, and
splitting the \code{Final.SNPs.LEO.NB.OCA} column into a vector of SNPs for
the target and source nodes respectively, and a second table with SNP to Node
relationships.
}

