% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_print_plot.R
\name{plot.structure.impact}
\alias{plot.structure.impact}
\title{Plot "network.impact" objects}
\usage{
\method{plot}{structure.impact}(x, order = c("alphabetical", "value"),
  zscore = FALSE, ...)
}
\arguments{
\item{x}{an output object from an impact function (class \code{network.impact})}

\item{order}{"alphabetical" orders nodes alphabetically, "value" orders nodes from
highest to lowest impact value}

\item{zscore}{logical. Converts raw impact statistics to z-scores for plotting}

\item{...}{other plotting specifications (ggplot2)}
}
\description{
Convenience function for generating network structure impact plots
}
\details{
Inputting an object of class \code{network.impact}
will return a line plot that shows the relative network impacts of each node.
}
\examples{
out1 <- global.impact(depression)
plot(out1)
plot(out1, order="value", zscore=TRUE)
out2 <- impact(depression)
plot(out2$Network.Structure)
}
