% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{social}
\alias{social}
\title{Simulated Social Engagement Data}
\format{
a dataframe. Columns represent individuals (nodes) and rows represent engagement in social media group conversations
}
\usage{
social
}
\description{
This simulated dataset contains binary social engagement scores for 16 individuals. For 400 social media posts
on a group forum, individuals were given a score of 1 if they engaged in group conversation regarding the post,
and a score of 0 if they did not engage with the post.
}
\examples{
head(social)
}
\keyword{datasets}
