% Generated by roxygen2 (4.0.0): do not edit by hand
\name{filterNEO}
\alias{filterNEO}
\title{Filter a NEO edge table by the recommended thresholds}
\usage{
filterNEO(dt)
}
\arguments{
\item{edgeTable}{edge.table generated by \code{\link{parseNEOcsv}}.}
}
\value{
a subset of the edge table containing only significant edges.
}
\description{
Filter a NEO edge table by the recommended thresholds
}

