% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktools-package.r
\docType{package}
\name{networktools-package}
\alias{networktools}
\alias{networktools-package}
\title{networktools: Tools for Identifying Important Nodes in Networks}
\description{
Includes assorted tools for network analysis. Bridge centrality; goldbricker; MDS, PCA, & eigenmodel network plotting.
}
\references{
Cite this package: \href{https://scholar.google.com/citations?view_op=view_citation&hl=en&user=KRBoSPoAAAAJ&citation_for_view=KRBoSPoAAAAJ:UeHWp8X0CEIC}{Google Scholar}

Cite bridge centrality: \href{https://www.paytonjjones.com/publication/bridge_centrality/jones_ma_mcnally_2019.pdf}{PDF}

Cite goldbricker / node redundancy: \href{https://pubmed.ncbi.nlm.nih.gov/30102777/}{PubMed}

Cite plots: \href{https://www.frontiersin.org/articles/10.3389/fpsyg.2018.01742/full}{Frontiers}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=networktools}
  \item Report bugs at \url{https://github.com/paytonjjones/networktools/issues}
}

}
\author{
\strong{Maintainer}: Payton Jones \email{paytonjjones@gmail.com}

}
\keyword{internal}
