\docType{package}
\name{networkTools-package}
\alias{networkTools}
\alias{networkTools-package}
\title{Tools for analysing and visualising networks.}
\description{
  This R package provides functions for analysisng and
  visualising large networks, primarily those generated
  from the R package WGCNA.
}
\details{
  \tabular{ll}{ Package: \tab networkTools\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2014-02-19\cr
  License: \tab GPL (>= 2)\cr }
}
\author{
  Scott Ritchie \email{sritchie73@gmail.com}
}

