% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktools-package.r
\docType{package}
\name{networktools}
\alias{networktools}
\title{networktools.}
\description{
Tools for Identifying Important Nodes in Networks
}
\details{
Includes assorted tools for network analysis. Bridge centrality (\code{bridge()})
calculates the extent to which nodes act as important communication points between
predefined community structures. The \code{goldbricker()} and \code{net_reduce()} functions
are tools for dealing with redundant nodes. For plotting networks, see \code{MDSnet()},
\code{PCAnet()}, and \code{EIGENnet()}.

For a complete list of functions, use library(help = "networktools")

For a complete list of vignettes, use browseVignettes("networktools")

Scholarly references:

Bridge centrality: https://www.paytonjjones.com/publication/bridge_centrality/jones_ma_mcnally_2019.pdf
Node redundancy: https://pubmed.ncbi.nlm.nih.gov/30102777/
Interpretable network plotting: https://www.frontiersin.org/articles/10.3389/fpsyg.2018.01742/full
}
\author{
Payton J. Jones
}
