% Generated by roxygen2 (4.0.0): do not edit by hand
\name{PlotNetworkHeatmap}
\alias{PlotNetworkHeatmap}
\title{Plot network edge weights as a heatmap.

Plots the edge weights between nodes as a heatmap, optionally drawing boxes
around pre-determined clusters. The \code{xlab} and \code{ylab} arguments are
useful when comparing the edge weights.}
\usage{
PlotNetworkHeatmap(edge.matrix, network.labels = NULL, xlab = "",
  ylab = "", legend.main = "", cluster.legend.main = "",
  heatmap.bins = 11, edge.weight.range = c(-1, 1), heatmap.gradient,
  cluster.cols)
}
\arguments{
\item{edge.matrix}{a square matrix containing the edge weights between each
pair of nodes.}

\item{network.labels}{an optional named vector assigning each node to a
network/cluster.}

\item{xlab}{x-axis heatmap label.}

\item{ylab}{y-axis heatmap label.}

\item{legend.main}{title to give to the legend denoting heatmap colors.}

\item{cluster.legend.main}{title to give to the legend denoting cluster
assignment.}

\item{heatmap.bins}{Number of bins to partition edge weights into for
plotting.}

\item{edge.weight.range}{The range of values edge weights can take. Defaults
to \code{c(-1, 1)}.}

\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
the heatmap.bins.}

\item{cluster.cols}{an optional vector of colors for denoting each cluster
on a legend.}
}
\description{
Plot network edge weights as a heatmap.

Plots the edge weights between nodes as a heatmap, optionally drawing boxes
around pre-determined clusters. The \code{xlab} and \code{ylab} arguments are
useful when comparing the edge weights.
}
\note{
It expects nodes to be pre-ordered within the edge weight matrix:
 I.e. clusters should be in continuous blocks. Otherwise, garbage-in,
 garbage-out.
}

